Removed Duplicate Code activating addStudent Dialog, kept working button in 2 places, added test ids.
