+++ src/components/AddStudent.js
@@ -18,83 +18,23 @@ import ButtonGroup from '@mui/material/ButtonGroup';
 import AppBar from '@mui/material/AppBar';
 import Toolbar from '@mui/material/Toolbar';
 import Typography from '@mui/material/Typography';
-
+import AddStudentButton from './AddStudentButton';
 toast.configure();
 
 // properties addStudent is required, function called when Add clicked.
 class AddStudent extends Component{
     constructor(props) {
 		  super(props);
-		  this.state = {open: false
-		  , student:{ 
-			  email: ''
-			  ,name: ''
-			  ,status: ''
-			  ,statusCode: -1
-			  ,student_id: -1
-		}
-		  ,sname:'',semail:'',sstatus:'',sstatusCode:'',sstudent_id:-1 };
+		  this.state = { };
 		 // console.log('line 27 constructor added addStudent with state:',this.state.student);
     };
-    handleClickOpen = () => {
-      this.setState( {open:true} );
-    };
-    handleClose = () => {
-      this.setState( {open:false} );
-	  //console.log('Line 45 AddStudent.js says: student, sstatusCode: ',this.state.student,this.state.sstatusCode);
-    };
-	setProp = (name,val)=>{
-		//console.log('Line 48 AddStudent.js says: name,val:',name,val);
-		switch(name){
-			case 'student_name':
-				this.setState({student:{name: val}});
-				this.setState({sname:val});
-				break;
-			case 'student_email':
-				this.setState({student:{email: val}});
-				this.setState({semail:val});
-				break;
-			case 'student_status':
-				this.setState({student:{status: val}});
-				this.setState({sstatus:val});
-				break;
-			case 'student_statusCode':
-				this.setState({student:{statusCode: val}});
-				this.setState({sstatusCode:val});
-				//console.log('Line 65 AddStudent.js says: ',this.state.student,this.state.student,this.state.sname,this.state.semail,this.state.sstatusCode);
-				break;
-		}
-		//console.log('Line 68 setProp.js says: ',this.state.student,this.state.sname,this.state.semail,this.state.sstatus,this.state.sstatusCode);
-	};
-    handleChange = (event) => {
-		//console.log('Line 72 AddStudent.js says: ',event.target.value,event.target.name);
-		this .setProp(event.target.name,event.target.value);
-		//console.log('Line 74 AddStudent.js says: ',this.state.student,this.state.student,this.state.sname,this.state.semail,this.state.sstatusCode);
-		//this.setState({student:{name:this.state.sname,email:this.state.semail,status:this.state.sstatus,statusCode:this.state.sstatusCode}});
-		//it's threaded, you can't do that
-		//console.log('Line 76 AddStudent.js says: ',this.state.student,this.state.sname,this.state.semail,this.state.sstatus,this.state.sstatusCode);
-    }
-  // Save student and close modal form
-    handleAdd = () => {
-		if ( ! (/^[0-9]+$/.test(this.state.sa_statusCode) ) ) {  
-			this.setState({message: 'The status code needs to be an integer'});
-			toast.error("The status code needs to be an integer", {
-				position: toast.POSITION.BOTTOM_LEFT
-			});
-			return;
-		}
-       // this.props.addStudent(this.state.student);
-	   this.setState({student:{name:this.state.sname,email:this.state.semail,status:this.state.sstatus,statusCode:this.state.sstatusCode}},this.addStudent);
-       // this.addStudent(this.state.student);
-       this.handleClose();
-    }
     render()  { 
       return (
           <div>
 		    <AppBar position="static" color="default">
             <Toolbar>
                <Typography variant="h7" color="inherit">
-                  { 'Student ' + this.props.location.name + ' ' +this.props.location.email  + ' ' +JSON.stringify(this.props.location)+ ' ' +JSON.stringify(this.props.location.student) }
+                  { 'Student '+JSON.stringify(this.props.location)+ ' ' +JSON.stringify(this.props.location.student) }
                 </Typography>
             </Toolbar>
           </AppBar>
@@ -103,120 +43,12 @@ class AddStudent extends Component{
                 For DEBUG:  display state.
                 {JSON.stringify(this.state)}
             </div>			
-            <Button variant="outlined" color="primary" style={{margin: 10}} onClick={this.handleClickOpen}>
-              Add Student
-            </Button>
-            <Dialog open={this.state.open} onClose={this.handleClose}>
-                <DialogTitle>Add Student</DialogTitle>
-                <DialogContent  style={{paddingTop: 20}} >
-                  <TextField autoFocus fullWidth label="Student Name" name="student_name" onChange={this.handleChange}  /> 
-                  <TextField fullWidth label="Student Email" name="student_email" onChange={this.handleChange}  /> 
-                  <TextField fullWidth label="Student Status" name="student_status" onChange={this.handleChange}  /> 
-                  <TextField fullWidth label="Student Status Code" name="student_statusCode" onChange={this.handleChange}  /> 
-                </DialogContent>
-                <DialogActions>
-                  <Button color="secondary" onClick={this.handleClose}>Cancel</Button>
-                  <Button id="Add" color="primary" onClick={this.handleAdd}>Add</Button>
-                </DialogActions>
-              </Dialog>
-			  <ToastContainer />
+            <AddStudentButton />
+			<ToastContainer />
           </div>
           </div>
       ); 
     }
-  // Add student
-  // addStudent = (student) => {
-  addStudent = () => {
-	  let student=this.state.student	  
-	  //const {student}=this.state.student;
-	  //console.log('102');
-    const token = Cookies.get('XSRF-TOKEN');
-	console.log('line 103 of AddStudent.js says student is:',student);
-    fetch(`${SERVER_URL}/student`,
-      { 
-        method: 'POST', 
-        headers: { 'Content-Type': 'application/json',
-                   'X-XSRF-TOKEN': token  }, 
-        body: JSON.stringify(student)
-      })
-    .then(res => {
-        if (res.ok) {
-						//console.log('120 add student res:',res);
-
-          toast.success("Student successfully added", {
-              position: toast.POSITION.BOTTOM_LEFT
-          });
-          //this.fetchStudent();
-		  
-		  //console.log('line 127 this is good res.json():',res.json())
-		  
-        } else {
-          toast.error("Error when adding", {
-              position: toast.POSITION.BOTTOM_LEFT
-          });
-          console.error('Post http status =' + res.status);
-        }return res;})
-		
-		
-		/*straight from Game.js:*/
-		  .then(res =>res.json())
-      .then(res => {
-		  //console.log('line 140 this is good res:',res);
-       this.setState({
-			student:{
-			  email: res.email
-			  ,name: res.name
-			  ,status: res.status
-			  ,statusCode: res.statusCode
-			  ,student_id: res.student_id}
-          ,semail: res.email
-          ,sname: res.name
-          ,sstatus: res.status
-          ,sstatusCode: res.statusCode
-          ,sstudent_id: res.student_id
-       });
-	   toast.success("DB row successfully added in student table.", {
-              position: toast.POSITION.BOTTOM_LEFT
-          });
-      })
-    .catch(err => {
-      toast.error("Error when adding", {
-            position: toast.POSITION.BOTTOM_LEFT
-        });
-        console.error(err);
-    })
-  } 
-  fetchStudent = () => {
-    //console.log("Line 90 AddStudent.fetchStudent");
-    const token = Cookies.get('XSRF-TOKEN');
-    
-    fetch(`${SERVER_URL}/student?id=${this.props.student.id}`, 
-      {  
-        method: 'GET', 
-        headers: { 'X-XSRF-TOKEN': token }
-      } )
-    .then((response) => {
-      console.log("FETCH RESP:"+response);
-      return response.json();}) 
-    .then((responseData) => { 
-      // do a sanity check on response
-      if (Array.isArray(responseData.student)) {
-        this.setState({ 
-          student: responseData.student,
-        });
-      } else {
-        toast.error("Fetch failed.", {
-          position: toast.POSITION.BOTTOM_LEFT
-        });
-      }        
-    })
-    .catch(err => {
-      toast.error("Fetch failed.", {
-          position: toast.POSITION.BOTTOM_LEFT
-        });
-        console.error(err); 
-    })
-  }
 }
 // required property:  addStudent is a function to call to perform the Add action
 AddStudent.propTypes = {
